+++ Assets/Picola.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c42e261095f53224ba811dc95b5bca84
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
