+++ Assets/Picola/PicoChan.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bc870fca26af84547988ad6a146c73ee
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
