+++ Assets/Picola/PicoChan/URP-2020.3+.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 84b482bef41808a40ab8db235f9435ca
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
