+++ Assets/Scripts/Audio/AudioEvents_Music.cs_X.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b670f08bdb6946e42bd590d622194419
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
