--- Assets/Scripts/Audio/FMOD_AudioEvents.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ca82b4e403ae4c94ba55fb3d4459b6b2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
